Make package work with isolated build
When using `catkin_make_isolated` or `catkin build`, the
`${CMAKE_SOURCE_DIR}` does not refer to the workspace root so this
cannot be used to construct the path to the pybind11 package in a
portable way.
As a workaround for now, specify the relative path from the root of this
package (determined via `${PROJECT_SOURCE_DIR}`).

Since setting `CMAKE_MODULE_PATH` for the parent scope won't work
either, use a template file which will get the absolute path set during
build and use this for CFG_EXTRAS.

Remove the `add_subdirectory` call which does not seem to be necessary.